move hostlist code into internal folder
